requirements.txt
absl-py==0.12.0
aio-pika==6.8.2
aiofiles==0.8.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bidict==0.22.0
boto3==1.24.2
botocore==1.27.2
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cloudpickle==1.6.0
colorama==0.4.4
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==37.0.2
cycler==0.11.0
decorator==4.4.2
dm-tree==0.1.7
dnspython==1.16.0
docopt==0.6.2
executing==0.8.3
fbmessenger==6.0.0
future==0.18.2
gast==0.3.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.20.4
httptools==0.4.0
httpx==0.15.4
humanfriendly==10.0
idna==2.10
importlib-metadata==4.11.4
ipython==8.4.0
jedi==0.18.1
jmespath==1.0.0
joblib==0.15.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.3.4
matplotlib-inline==0.1.3
mattermostwrapper==2.2
multidict==4.7.6
networkx==2.5.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==2.0.10
protobuf==3.20.1
psycopg2-binary==2.8.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
Pygments==2.12.0
PyJWT==2.4.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.18.1
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2021.3
questionary==1.9.0
rasa==2.6.2
rasa-sdk==2.6.0
redis==3.5.3
regex==2020.9.27
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rocketchat-API==1.15.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.5.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.2
scipy==1.8.1
sentry-sdk==0.19.5
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
sniffio==1.2.0
SQLAlchemy==1.3.24
stack-data==0.2.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.4
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
termcolor==1.1.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tqdm==4.59.0
traitlets==5.2.2.post1
twilio==6.50.1
typeguard==2.13.3
typing-extensions==3.10.0.2
tzlocal==2.1
ujson==4.3.0
urllib3==1.26.9
wcwidth==0.2.5
webexteamssdk==1.6
websockets==8.1
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
